@@ -5,20 +5,22 @@ author: Valery Colong<valery.colong@activspaces.com>
 homepage: https://github.com/valerycolong/vcard
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  validators: ^2.0.0+1
-  path_provider: ^1.6.5
-  path: ^1.6.4
-  intl: ^0.16.0
+  validators: ^3.0.0
+  path_provider: ^2.0.8
+  path: ^1.8.0
+  intl: ^0.17.0
   meta: ^1.1.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
